Remove JWTCallerPrinciple abstraction
